External extensions tweaks (#203)
