+++ .changeset/mighty-singers-burn.md
@@ -1,0 +1,5 @@
+---
+"create-eth": patch
+---
+
+Add trusted GitHub organizations for extensions
